requirements.txt
fastapi
uvicorn
pandas
opencv-python-headless
numpy
Pillow
torch
yolov5 @ git+https://github.com/ultralytics/yolov5.git
